rename network -> nodes
