+++ chain-signatures/node/src/metrics/mod.rs
@@ -5,7 +5,7 @@ use prometheus::{HistogramOpts, HistogramVec, Opts, Result};
 pub mod hardware;
 pub mod indexers;
 pub mod messaging;
-pub mod network;
+pub mod nodes;
 pub mod protocols;
 pub mod requests;
 pub mod storage;
